releases/v--help/scripts/requirements.txt
requests
python-dotenv
pytest
chardet